some example about spring transaction

+++ java/spring/transaction/src/main/java/com/example/transaction/xml/service/EmployeeService.java
@@ -3,8 +3,8 @@ package com.example.transaction.xml.service;
 import com.example.transaction.xml.data.Employee;
 
 public interface EmployeeService {
-    Employee getEmployee(Integer id);
+    Employee getEmployee(Long id);
     void insertEmployee(Employee employee);
     void updateEmployee(Employee employee);
-    void deleteEmployee(Integer id);
+    void deleteEmployee(Long id);
 }
